Move step to `protocol` folder
