+++ src/lib.rs
@@ -1,13 +1,13 @@
 #![deny(clippy::clone_on_ref_ptr)]
 
 mod chunkscan;
-mod circuit;
 #[cfg(feature = "cli")]
 pub mod cli;
 pub mod error;
 pub mod field;
 pub mod helpers;
 pub mod net;
+pub mod protocol;
 pub mod prss;
 mod replicated_secret_sharing;
 pub mod securemul;
